Fix WearBarParams (de)serialization crash due to unsatisfied interface
While image/color.NRGBA has a (de)serialization override, it does not
satisfy the (de)serializer interface. This causes a crash when certain
item definitions are sent or received. This commit fixes the issue by
manually copying the (de)serialization override of image/color.NRGBA to
the WearBarParams override implementation instead of calling the
(de)serialize function that requires the input to satisfy
(de)serializer.